Trying to get textures to work.

+++ src/terrain.cpp
@@ -43,8 +43,8 @@ void Terrain::generate_height_map(GLuint size, GLfloat max_height, GLint village
 	//Offset of Village on Island, Round Plateau.
 	float r_ratio_offset = 0.5; //Row <----Currnently, Code only works for when village is at center of island.
 	float c_ratio_offset = 0.5; //Col
-	unsigned int village_mid_r = size * r_ratio_offset;
-	unsigned int village_mid_c = size * c_ratio_offset;
+	unsigned int village_mid_r = (unsigned int) size * r_ratio_offset;
+	unsigned int village_mid_c = (unsigned int) size * c_ratio_offset;
 
 	unsigned int radius = village_diameter / 2;
